parsegraph-back-atlantis/requirements.txt
aniso8601==9.0.1
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
flask==3.0.2
flask-restx==1.3.0
flask-sqlalchemy==3.1.1
fonttools==4.49.0
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.9
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.27
typing-extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.17.0
